requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
distlib==0.3.8
Django==3.2.23
django-crispy-forms==2.0
django-quill-editor==0.1.40
filelock==3.12.2
importlib-metadata==6.7.0
Pillow==9.5.0
platformdirs==4.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.7.1
zipp==3.15.0
